requirements.txt
sqlalchemy
psycopg2
pyyaml
flask
flask-restful
flask-sqlalchemy
flask-cors
